backend/requirements.txt
# FastAPI Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP Client
httpx>=0.26.0

# SSE (Server-Sent Events)
sse-starlette>=1.8.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async Support
anyio>=4.2.0

# ML Dependencies
numpy>=1.26.0
scikit-learn>=1.4.0

# Future ML Dependencies (commented out for now)
# sentence-transformers>=2.2.0
# faiss-cpu>=1.7.0
# chromadb>=0.4.0
